Consolidate 6+ comment components into CommentDialog + ReplyDialog
Replace answering-dialog, update-comment-dialog, and scattered local
state with two URL-driven dialog components. All dialog activation is
now derived from URL search params, not useState. Delete the old
components and rewire comment-with-replies to use Link-based edit
and reply buttons.
